+++ lib/src/properties/text_field_props.dart
@@ -1,7 +1,5 @@
 import 'dart:ui' as ui show BoxHeightStyle, BoxWidthStyle;
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/gestures.dart';